requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
pandas
python-dotenv
streamlit